The Timestamp Type Patch. Added a timestamp type for boottime, etc. The new type helps format metrics on the webpage, among other things.
Other changes include new OWNER, LATLONG attributes to the CLUSTER tag. The
values for these metrics are set in /etc/gmond.conf. The OWNER attr ensures
that the cluster will have a unique name accross admin domains. The NAME/OWNER
pair should be unique for every cluster in the world. This helps for grid
management, and is required by Globus.

The LATLONG attr specifies a cluster's location by its latitude and longitude.
It is not strictly necessary, but enables some cool analysis and perhaps even
a world map with clusters on it for the webpage. I have chosen to use the
DMS format like "N32.87 W117.22", which will achieve around a 1 mile accuracy.

Smaller changes include renaming the units "KBs" to "KB", and began work on
adding a LOCATION attribute to the HOST tag in the XML output.

-Federico


git-svn-id: http://ganglia.svn.sourceforge.net/svnroot/ganglia/trunk/monitor-core@46 93a4e39c-3214-0410-bb16-828d8e3bcd0f