The grid-aware patch.
Gmetad now wraps its data sources in a GRID tag. Nested GRIDs are allowed.
The purpose for this new tag is to distribute the load of keeping RRDs, which
has been identified as a computation and I/O bottleneck.

New features:
-Grid Composability (meta-meta*)
-Can use hostnames in trusted host declarations
-Settable polling interval for sources [1]
-Debug command line option
-Cleaner code for rrd_helpers.c and others.
-Keeps swap_total summary RRDs for swap display on MEM graphs.

[1] The polling interval is the time between XML pulls from a data source.
It is optional, with a default of 15sec as before. Be aware that the
RRDs rely on this value when deciding if a primary data point (PDP) is
"undefined". You should only change this value if you are creating new
RRDs. If the polling interval is changed after the RRD files have been
created, graphs in the webfrontend will not display correctly. Perhaps
this can be changed in the future with a call to "rrdtune".

This code has been tested on Linux only, and is not ready for release until
verified on other platforms.

Federico


git-svn-id: http://ganglia.svn.sourceforge.net/svnroot/ganglia/trunk/monitor-core@166 93a4e39c-3214-0410-bb16-828d8e3bcd0f