build: linux: fix biarch support
remove unreliable check for host_cpu which will backfire in powerpc64 where
binaries are expect to be build 32-bit by default as reported by Jarod Wilson
while building fedora packages for ppc32 and ppc64.

use instead an autoconf compile test and file to detect which ABI we are
building for to try to guess the right libdir to use for it.


git-svn-id: http://ganglia.svn.sourceforge.net/svnroot/ganglia/trunk/monitor-core@1468 93a4e39c-3214-0410-bb16-828d8e3bcd0f