gmond: reformat header files and definitions for clarity
partially reverts r1489 to implicitally include stdlib.h which was added
in r1040 and is still needed for *rand() and atoi().

add strings.h for strcasecmp() and time.h for time() and add documentation
and visual aids so it is clear how to use this as a template for building
custom metric modules.


git-svn-id: http://ganglia.svn.sourceforge.net/svnroot/ganglia/trunk/monitor-core@1513 93a4e39c-3214-0410-bb16-828d8e3bcd0f