gmond: refactor language directive for compiled DSO
partially revert r1500 re-adding the language directive for compiled DSOs for
consistency as suggested by Brad.

another attempt at documenting the changes in a language agnostic way so that
the use of language="C" can be better understood as meaning "C" bindings that
are compatible with all "C" like languages and used by any compiled DSO
regardless of what language was used to code it.


git-svn-id: http://ganglia.svn.sourceforge.net/svnroot/ganglia/trunk/monitor-core@1530 93a4e39c-3214-0410-bb16-828d8e3bcd0f