documentation: update/improve instructions for building under cygwin
As suggested by Daniel Pocock, include the group where the dependencies
can be found if using the tree view.  Update names of libraries that had
changed since and reflect newer versions of libraries when provided.

Avoid a libconfuse bug triggered by the NLS support and that will be
hopefully fixed in a future version but that the current workarounds
don't catch at configure time.


git-svn-id: http://ganglia.svn.sourceforge.net/svnroot/ganglia/trunk/monitor-core@1943 93a4e39c-3214-0410-bb16-828d8e3bcd0f